refactor(dynasty): extract common request list logic into RequestList component
