+++ src/Pages/Profile/Tabs/dynasty/components/RequestList.jsx
@@ -1,0 +1,191 @@
+import "react-multi-date-picker/styles/backgrounds/bg-dark.css";
+import "react-multi-date-picker/styles/colors/yellow.css";
+
+import DatePicker from "react-multi-date-picker";
+import { FaRegCalendarAlt } from "react-icons/fa";
+import TimePicker from "react-multi-date-picker/plugins/time_picker";
+import persian from "react-date-object/calendars/persian";
+import persian_fa from "react-date-object/locales/persian_fa";
+import styled from "styled-components";
+import { useState, useEffect } from "react";
+import Title from "../../../../../Components/Title";
+import SearchInput from "../../../../../Components/SearchInput";
+import useRequest from "../../../../../Services/Hooks/useRequest";
+import { getFieldTranslationByNames } from "../../../../../Services/Utility";
+
+const Container = styled.div`
+  padding: 20px 15px 0px 0;
+  overflow-y: auto;
+  height: 550px;
+  @media (min-width: 640px) { height: 290px; }
+  @media (min-width: 740px) { height: 270px; }
+  @media (min-width: 840px) { height: 290px; }
+  @media (min-width: 890px) { height: 315px; }
+  @media (min-width: 930px) { height: 330px; }
+  @media (min-width: 1024px) { height: 400px; }
+  @media (min-width: 1180px) { height: 600px; }
+  @media (min-width: 1280px) { max-height: 945px; }
+  @media (min-width: 1366px) { height: 650px; }
+  @media (min-width: 1920px) { height: 640px; }
+`;
+
+const Div = styled.div`
+  display: grid;
+  grid-template-columns: 3fr 2fr;
+  align-items: center;
+  gap: 20px;
+  margin-top: 15px;
+  @media (min-width: 1024px) {
+    grid-template-columns: 3fr 1fr;
+  }
+`;
+
+const Date = styled.div`
+  border-radius: 5px;
+  border: 1px solid #454545;
+  display: flex;
+  flex-grow: 1;
+  align-items: center;
+  justify-content: space-between;
+  overflow: hidden;
+  background-color: ${(props) => props.theme.colors.newColors.otherColors.inputBg};
+  height: 48px;
+  padding: 0 10px;
+  svg {
+    color: ${(props) => props.theme.colors.newColors.shades.title};
+  }
+  input {
+    color: ${(props) => props.theme.colors.newColors.shades.title};
+    width: 100%;
+    height: 100%;
+    background-color: transparent;
+    border: none;
+    outline: none;
+    &:focus {
+      border: none;
+      outline: none;
+    }
+    font-size: 16px;
+  }
+`;
+
+const RequestList = ({ 
+  title,
+  requestType,
+  ListComponent,
+  userField = "to_user" // برای sent list از to_user و برای received list از from_user استفاده می‌شود
+}) => {
+  const { Request } = useRequest();
+  const [searched, setSearched] = useState("");
+  const [rows, setRows] = useState([]);
+  const [status, setStatus] = useState({
+    pending: false,
+    confirmed: false,
+    failed: false,
+  });
+
+  const [member, setMember] = useState({
+    father: false,
+    mother: false,
+    child: false,
+    sister: false,
+    brother: false,
+    wife: false,
+  });
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await Request(`dynasty/requests/${requestType}`);
+        const formattedData = response.data.data.map((item) => ({
+          id: item.id,
+          code: item[userField].code,
+          date: item.date,
+          time: item.time,
+          member: item.relationship,
+          status: item.status === 1 ? "confirmed" : item.status === 0 ? "pending" : "failed",
+          member_slug: getMemberSlug(item.relationship),
+          gif: item?.prize?.satisfaction || 0,
+          psc: item?.prize?.psc || 0,
+        }));
+        setRows(formattedData);
+      } catch (error) {
+        console.error("Error fetching data:", error);
+      }
+    };
+
+    fetchData();
+  }, [requestType, userField]);
+
+  const getMemberSlug = (relationship) => {
+    const slugMap = {
+      برادر: "brother",
+      خواهر: "sister",
+      پدر: "father",
+      مادر: "mother",
+      فرزند: "child",
+      همسر: "wife",
+    };
+    return slugMap[relationship] || relationship.toLowerCase();
+  };
+
+  const filteredItems = rows.filter((row) => {
+    const codeMatch = row.code.toString().includes(searched);
+    const statusMatch =
+      (!status.confirmed && !status.failed && !status.pending) ||
+      (status.confirmed && row.status === "confirmed") ||
+      (status.failed && row.status === "failed") ||
+      (status.pending && row.status === "pending");
+    const memberMatch =
+      (!member.child &&
+        !member.wife &&
+        !member.brother &&
+        !member.sister &&
+        !member.father &&
+        !member.mother) ||
+      (member.child && row.member_slug === "child") ||
+      (member.wife && row.member_slug === "wife") ||
+      (member.brother && row.member_slug === "brother") ||
+      (member.sister && row.member_slug === "sister") ||
+      (member.father && row.member_slug === "father") ||
+      (member.mother && row.member_slug === "mother");
+
+    return codeMatch && statusMatch && memberMatch;
+  });
+
+  return (
+    <Container>
+      <div>
+        <Title title={getFieldTranslationByNames(title)} />
+      </div>
+      <Div>
+        <SearchInput
+          onchange={(e) => setSearched(e.target.value)}
+          value={searched}
+          placeholder={getFieldTranslationByNames(849)}
+        />
+        <Date>
+          <DatePicker
+            placeholder={getFieldTranslationByNames(564)}
+            className="bg-dark yellow"
+            format="YYYY/DD/MM HH:mm:ss"
+            plugins={[<TimePicker position="bottom" />]}
+            calendar={persian}
+            locale={persian_fa}
+            calendarPosition="bottom-right"
+          />
+          <FaRegCalendarAlt size={20} />
+        </Date>
+      </Div>
+      <ListComponent
+        setStatus={setStatus}
+        setMember={setMember}
+        member={member}
+        status={status}
+        rows={filteredItems}
+      />
+    </Container>
+  );
+};
+
+export default RequestList;
